requirements.txt
Flask==3.1.0
Flask-SocketIO==5.5.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
matplotlib==3.10.1
reportlab==4.3.1
simple-websocket==1.1.0
waitress==3.0.2
